feat: move toggle privacy button to app bar
