+++ frontend/app/src/components/UserDropdown.vue
@@ -35,47 +35,6 @@
             {{ $t('user_dropdown.settings') }}
           </v-list-item-title>
         </v-list-item>
-        <v-menu
-          offset-x
-          left
-          :max-width="400"
-          :close-on-content-click="false"
-          content-class="user-dropdown__privacy-mode__menu"
-        >
-          <template #activator="{ on, attrs }">
-            <v-list-item
-              key="privacy-mode"
-              data-cy="user-dropdown__privacy-mode"
-              class="user-dropdown__privacy-mode"
-              v-bind="attrs"
-              v-on="on"
-            >
-              <v-list-item-avatar>
-                <v-icon color="primary">{{ privacyModeIcon }}</v-icon>
-              </v-list-item-avatar>
-              <v-list-item-title>
-                {{ $t('user_dropdown.change_privacy_mode.label') }}
-              </v-list-item-title>
-            </v-list-item>
-          </template>
-          <v-card>
-            <div class="slider-wrapper pa-8" :style="sliderWrapperStyle">
-              <v-slider
-                :value="privacyMode"
-                data-cy="user-dropdown__privacy-mode__input"
-                :class="`selected-${privacyMode}`"
-                :tick-labels="tickLabels"
-                :step="1"
-                :max="2"
-                :min="0"
-                ticks="always"
-                :tick-size="4"
-                vertical
-                @change="changePrivacyMode($event)"
-              />
-            </div>
-          </v-card>
-        </v-menu>
 
         <v-divider class="mx-4" />
         <v-list-item
@@ -103,7 +62,7 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, ref } from '@vue/composition-api';
+import { defineComponent, ref } from '@vue/composition-api';
 import { get, set } from '@vueuse/core';
 import ConfirmDialog from '@/components/dialogs/ConfirmDialog.vue';
 import MenuTooltipButton from '@/components/helper/MenuTooltipButton.vue';
@@ -111,29 +70,6 @@ import MenuTooltipButton from '@/components/helper/MenuTooltipButton.vue';
 import { useRoute, useRouter } from '@/composables/common';
 import { setupSession } from '@/composables/session';
 import { interop } from '@/electron-interop';
-import i18n from '@/i18n';
-
-const tickLabels: string[] = [
-  i18n.t('user_dropdown.change_privacy_mode.normal_mode.label').toString(),
-  i18n
-    .t('user_dropdown.change_privacy_mode.semi_private_mode.label')
-    .toString(),
-  i18n.t('user_dropdown.change_privacy_mode.private_mode.label').toString()
-];
-
-// TODO: Remove css variable for tick description in Vue3 and use State Driven Dynamic CSS
-// https://v3.vuejs.org/api/sfc-style.html#state-driven-dynamic-css
-const sliderWrapperStyle: { [key: string]: string } = {
-  '--tick-description-3': `'${i18n
-    .t('user_dropdown.change_privacy_mode.normal_mode.description')
-    .toString()}'`,
-  '--tick-description-2': `'${i18n
-    .t('user_dropdown.change_privacy_mode.semi_private_mode.description')
-    .toString()}'`,
-  '--tick-description-1': `'${i18n
-    .t('user_dropdown.change_privacy_mode.private_mode.description')
-    .toString()}'`
-};
 
 export default defineComponent({
   name: 'UserDropdown',
@@ -142,7 +78,7 @@ export default defineComponent({
     MenuTooltipButton
   },
   setup() {
-    const { username, privacyMode, changePrivacyMode, logout } = setupSession();
+    const { username, logout } = setupSession();
     const confirmLogout = ref<boolean>(false);
     const router = useRouter();
     const route = useRoute();
@@ -160,169 +96,11 @@ export default defineComponent({
       }
     };
 
-    const privacyModeIcon = computed<string>(() => {
-      return ['mdi-eye', 'mdi-eye-minus', 'mdi-eye-off'][get(privacyMode)];
-    });
-
     return {
       confirmLogout,
-      privacyModeIcon,
       username,
-      privacyMode,
-      changePrivacyMode,
-      tickLabels,
-      sliderWrapperStyle,
       logoutHandler
     };
   }
 });
 </script>
-
-<style scoped lang="scss">
-.user-dropdown {
-  &__privacy-mode {
-    &__menu {
-      .slider-wrapper {
-        .v-input {
-          ::v-deep {
-            .v-slider {
-              width: 210px;
-
-              &__track-container {
-                width: 4px;
-                left: auto !important;
-                right: 0;
-                transform: translateX(50%) !important;
-              }
-
-              &__track-background {
-                &.primary {
-                  background: var(--v-rotki-light-grey-darken1) !important;
-                }
-              }
-
-              &__ticks-container {
-                left: auto !important;
-                right: 0;
-                transform: translateX(50%) !important;
-              }
-
-              &__tick {
-                --color: var(--v-rotki-light-grey-darken2);
-                width: 10px !important;
-                height: 10px !important;
-                right: calc(50% - 5px) !important;
-                border-radius: 50%;
-                background-color: var(--color) !important;
-                cursor: pointer;
-
-                &-label {
-                  left: auto !important;
-                  right: 30px;
-                  text-transform: uppercase;
-                  font-weight: bold;
-                  font-size: 0.75rem;
-                  color: var(--color);
-                  text-align: right;
-                  margin-top: 5px;
-
-                  &:after {
-                    text-transform: none;
-                    font-weight: normal;
-                    display: block;
-                  }
-                }
-
-                @for $i from 1 through 3 {
-                  &:nth-child(#{$i}) {
-                    bottom: calc(150% - (50% * #{$i}) - 5px) !important;
-
-                    /* stylelint-disable selector-class-pattern, rule-empty-line-before */
-                    .v-slider {
-                      &__tick-label {
-                        &::after {
-                          content: var(--tick-description-#{$i});
-                        }
-                      }
-                    }
-                    /* stylelint-enable selector-class-pattern, rule-empty-line-before */
-                  }
-                }
-              }
-
-              &__thumb-container {
-                left: auto !important;
-                right: 0;
-              }
-
-              &__thumb {
-                &::before {
-                  transform: scale(1) !important;
-                }
-              }
-            }
-          }
-
-          /* stylelint-disable plugin/stylelint-bem-namics, rule-empty-line-before */
-          @for $i from 0 through 2 {
-            &.selected-#{$i} {
-              ::v-deep {
-                .v-slider {
-                  &__tick {
-                    @for $j from 3-$i through 3 {
-                      &:nth-child(#{$j}) {
-                        --color: var(--v-primary-base);
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-          /* stylelint-enable plugin/stylelint-bem-namics, rule-empty-line-before */
-        }
-      }
-    }
-  }
-}
-
-.theme {
-  &--dark {
-    .user-dropdown {
-      &__privacy-mode {
-        &__menu {
-          .slider-wrapper {
-            .v-input {
-              ::v-deep {
-                .v-slider {
-                  &__tick {
-                    --color: var(--v-secondary-lighten2);
-                  }
-                }
-              }
-
-              /* stylelint-disable plugin/stylelint-bem-namics, rule-empty-line-before */
-              @for $i from 0 through 2 {
-                &.selected-#{$i} {
-                  ::v-deep {
-                    .v-slider {
-                      &__tick {
-                        @for $j from 3-$i through 3 {
-                          &:nth-child(#{$j}) {
-                            --color: var(--v-primary-base);
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-              /* stylelint-enable plugin/stylelint-bem-namics, rule-empty-line-before */
-            }
-          }
-        }
-      }
-    }
-  }
-}
-</style>
