+++ frontend/app/tests/e2e/pages/rotki-app.ts
@@ -53,10 +53,9 @@ export class RotkiApp {
   }
 
   changePrivacyMode(mode: number) {
-    cy.get('.user-dropdown').click();
-    cy.get('[data-cy="user-dropdown__privacy-mode"]').click();
+    cy.get('.privacy-mode-dropdown').click();
     cy.get(
-      '[data-cy="user-dropdown__privacy-mode__input"] ~ .v-slider__thumb-container'
+      '[data-cy="privacy-mode-dropdown__input"] ~ .v-slider__thumb-container'
     ).as('input');
 
     cy.get('@input').focus().type('{downarrow}'.repeat(3));
@@ -64,7 +63,7 @@ export class RotkiApp {
     if (mode > 0) {
       cy.get('@input').type('{uparrow}'.repeat(mode));
     }
-    cy.get('[data-cy="user-dropdown__privacy-mode"]').click();
+    cy.get('.privacy-mode-dropdown').click();
     cy.get('.user-dropdown').click();
   }
